refactor(ble): split handler-esphome-proxy into module directory (#116)
Move the 695-line handler into src/ble/handler-esphome-proxy/{index,client,advert,scan,watcher}.ts with verbatim logic; add macToInt helper. Repoint src/ble/index.ts, src/runtime/sources.ts and the test import. No behavior change: 23 handler tests unchanged.
